backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
weaviate-client>=4.19.2
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
pydantic-ai>=1.58.0
logfire>=0.20.0
ddgs>=4.0.0
fake-useragent>=1.4.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
trafilatura>=1.6.0
xxhash>=3.4.0
tiktoken>=0.5.0

markdown-it-py>=3.0.0
tqdm>=4.66.0
networkx
yfiles-jupyter-graphs
